move hooks into own module
